+++ .gitignore
@@ -7,10 +7,14 @@ yarn-error.log*
 pnpm-debug.log*
 lerna-debug.log*
 
+
+coverage
 node_modules
+coverage
 dist
 dist-ssr
 *.local
+
 
 # Editor directories and files
 .vscode/*
