--- AuthHom/AuthHom_Pre-Build/src/Auth.css
@@ -1,56 +1,0 @@
-* {
-    font-family: 'Roboto', sans-serif;
-  }
-  
-  .authentication-container {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    margin: 1rem;
-    height: 50vh;
-    margin-bottom: 0rem;
-  }
-  
-  .authentication-header {
-    display: flex;
-    margin-bottom: 1rem;
-  }
-  
-  .authentication-header h1 {
-    text-align: center;
-    color: #ffffff;
-    font-size: 1.5rem;
-    font-weight: bold;
-    height: 2rem;
-    width: 12rem;
-    border-radius: 0.25rem;
-    background-color: #6c9a06;
-    padding: 0.5rem;
-  }
-  
-  form {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-  }
-  
-  label {
-    margin-top: 1rem;
-  }
-  
-  input {
-    padding: 0.5rem;
-    border-radius: 0.25rem;
-    border: 1px solid #ddd;
-    margin-bottom: 1rem;
-  }
-  
-  button {
-    padding: 0.5rem 1rem;
-    border-radius: 0.3rem;
-    border: none;
-    background-color: #0a351c;
-    color: #fff;
-    font-size: 1rem;
-    cursor: pointer;
-  }
