--- AuthHom/AuthHom_Pre-Build/src/Header.css
@@ -1,36 +1,0 @@
-.header-component {
-  display: flex;
-  background-color: #d9d9d9;
-  margin: 1.2rem;
-  padding: 0 1rem;
-  width: calc(100% - 2rem);
-  justify-content: space-between;
-  align-items: center;
-  height: 15vh;
-  box-sizing: border-box;
-  margin-bottom: 0rem;
-}
-
-.header-useravatar {
-  display: flex;
-  align-items: center;
-  font-size: 14px;
-}
-
-.header-useravatar img {
-  width: 40px;
-  height: 40px;
-  object-fit: cover;
-  border-radius: 50%;
-  margin-right: 8px;
-}
-
-.header-useravatar p {
-  margin: 0;
-}
-
-.header-component img {
-  max-height: 100%;
-  max-width: 100%;
-  object-fit: contain;
-}
