--- VersionChanges/VersionChanges_Build/VC_Build.txt
@@ -1,1 +1,0 @@
-Folder for the build of VersionChanges features
