--- VersionChanges/VersionChanges_Pre-Build/VersionChanges.txt
@@ -1,1 +1,0 @@
-New folder for VersionChanges related features
