+++ babel.config.cjs
@@ -1,0 +1,6 @@
+module.exports = {
+    presets: [
+      ['@babel/preset-env', { targets: { esmodules: true } }],
+      ['@babel/preset-react', { runtime: 'automatic' }]
+    ]
+}
