+++ jest.setup.js
@@ -1,0 +1,7 @@
+// En caso de necesitar la implementación del FetchAPI
+// npm i --save-dev isomorphic-fetch
+// o podemos instalar: (para versiones de node ≥18)
+// npm i --save-dev whatwg-fetch
+// import 'isomorphic-fetch'
+// o podemos instalar: (para versiones de node ≥18)
+// import 'whatwg-fetch'
