+++ sonar-project.properties
@@ -1,0 +1,15 @@
+sonar.projectKey=FabricaEscuela-Modulo5_Analisis1
+sonar.organization=fabricaescuela-modulo5
+
+sonar.javascript.lcov.reportPaths=./coverage/lcov.info
+
+sonar.coverage.exclusions=**/*test.js, **/*spec.js, **/*Test.js, **/*.cjs, /src/main.jsx
+sonar.test.exclusions=**/*test.js, **/*spec.js, **/*Test.js, **/*.cjs, /src/main.jsx
+# This is the name and version displayed in the SonarCloud UI.
+#sonar.projectName=CodeFactoryFrontTesting
+#sonar.projectVersion=1.0
+
+# Path is relative to the sonar-project.properties file. Replace "\" by "/" on Windows.
+# sonar.sources=.
+
+sonar.sourceEncoding=UTF-8
