+++ src/components/FooterC.jsx
@@ -1,0 +1,27 @@
+import '../Footer.css'
+function FooterC() {
+  return (
+    <footer className="app-footer">
+      <p>
+        Universidad de Antioquia | Vigilada Mineducación | Acreditación
+        institucional hasta el 2022 | NIT 890980040-8
+      </p>
+      <p>
+        Recepción de correspondencia: calle 70 No. 52 - 21 | Apartado Aéreo 1226
+        | Dirección: calle 67 No. 53 - 108 | Horario de atención
+      </p>
+      <p>
+        Conmutador: [57 + 604] 219 8332 | Línea gratuita de atención al
+        ciudadano: 018000 416384 | Fax: [57 + 604] 263 8282
+      </p>
+      <p>
+        Peticiones, quejas, reclamos, sugerencias, denuncias, consultas y
+        felicitaciones
+      </p>
+      <p>Política de tratamiento de datos personales</p>
+      <p>Medellín - Colombia | Todos los Derechos Reservados © 2020</p>
+    </footer>
+  );
+}
+
+export default FooterC;
